# Source/onespace-appdiscovery-service/src/main/resources/swagger/AppDiscoveryFramework.yaml
openapi: 3.0.3
info:
  title: OneSpace Platform APIs
  description: |-
    __OneSpace Platform APIs__ enables OneSpace Portal to register, list & approve Philips  serviceability applications.
  version: 1.0.0
servers:
  - url: https://apps.onespace-{region}-{env}.com/api
...
paths:
...
  /Application:
    post:
      tags:
        - Application APIs
      summary: Registers an application with the OneSpace Platform
      description: |
        A Session Cookie with 
        __EPO_APP.REGISTER__ permission is required to execute this operation.

      requestBody:
        description: Application registration details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
        required: true
      responses:
        '201':
          description: Application registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationID'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
    get:
      tags:
        - Application APIs
      summary: Lists applications entitled for the logged in user
      description: |
        A Session Cookie with 
        __EPO_APP.LIST__ permission is required to execute this operation.
        Search type supported is equals and case insensitive with AND multi-value search support.

      parameters:
        - name: status
          in: query
          description: The status of the application
          schema:
            type: string
            format: uuid
        - name: deployment
          in: query
          description: >-
            Setting the "deployment" criteria to a value equals to "multi-tenant" ensures that only multi-tenant applications are retrieved by this API. Applications that have deployment values such as per-region, instance etc. are not considered.
          schema:
            type: string
            example: 'multi-tenant, per-region, instance, not-applicable'
        - name: market
          in: query
          description: >-
            Setting the "market" criteria to a value equals to NAM, EU ensures that only applications that have been rolled out to NAM and EU markets are fetched.
          schema:
            type: string
            example: 'NAM, EU etc.'
        - name: region
          in: query
          description: >-
            An application could be rolled out in multiple regions such as NOR, DACH. The region parameter filters applications based on the region name"
          schema:
            type: string
            example: 'NOR, DACH etc.'
        - name: country
          in: query
          description: >-
            Country where the application available.
          schema:
            type: string
            example: 'United States, Canada etc.'
        - name: sub-region
          in: query
          description: >-
            Sub-region where the application available.
          schema:
            type: string
            example: 'Central, South, North etc.'
        - name: language
          in: query
          description: >-
            An application could support multiple languages. The Language parameter enables filtering of applications based on the specified language(s).
          schema:
            type: string
            example: 'Portuguese, Deutsch etc.'
        - name: sso-enabled
          in: query
          description: >-
            An application could support Single Sign On. The sso-enabled parameter enables filtering of applications based on specified value.
          schema:
            type: boolean
            example: true
        - name: specialty
          in: query
          description: >-
            An application could support multiple specialities. The speciality parameter enables filtering of applications based on the specified speciality.
          schema:
            type: string
            example: 'cardiology, oncology etc.'
        - name: modality
          in: query
          description: >-
            An application could support multiple modalities. The modality parameter enables filtering of applications based on the specified modality.
          schema:
            type: string
            example: 'ct, mr, ultrasound etc.'
        - name: category
          in: query
          description: >-
            An application could support multiple categories. The category parameter enables filtering of applications based on the specified category.
          schema:
            type: string
            example: 'admin app, business app etc.'

      responses:
        '200':
          description: Applications retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
...
components:
  schemas:
    Registration:
      type: object
      required:
        - name
      properties:
        id:
          description: Unique Identifier that represents an application in OneSpace platform
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          description: Name of the application
          type: string
          minLength: 3
          maxLength: 2000
          example: DoseWise Portal
        description:
          description: Description of the application
          type: string
          minLength: 3
          maxLength: 2000
          example: DoseWise Portal is a vendor-agnostic, web-based solution that collects, measures, analyzes, and reports patient & staff radiation exposure, assisting you to take control of quality of care, efficiency, patient and staff safety.
        version:
          description: Version of the application
          type: string
          minLength: 3
          maxLength: 50
          example: 1.0.0
        registeredBy:
          description: Email address of the application integrator requesting the application registration
          type: string
          minLength: 3
          maxLength: 120
          example: john.doe@philips.com
        registeredDateTime:
          type: string
          format: date-time
          description: Date & Time at which the application was approved by the approver
          example: 2024-01-22T13:09:04+0000
        isSSOEnabled:
          description: Is SSO Enabled
          type: boolean
          example: true
        isMFE:
          description: is this a microfrontend application
          type: boolean
          example: false
        languages:
          description: The languages supported by the applications
          type: array
          items:
            $ref: '#/components/schemas/Value'
          example: [
            { id: 330e8400-e29b-41d4-a716-446655440000, name: en-US},
            { id: 550e8400-e29b-41d4-a716-446655440000, name: pt-BR}]
        specialities:
          description: One or more medical specialties that this application is associated with
          type: array
          items:
            $ref: '#/components/schemas/Value'
          example: [
            { id: 330e8400-e29b-41d4-a716-446655440000, name: Radiology},
            { id: 550e8400-e29b-41d4-a716-446655440000, name: Cardiology}]
        modalities:
          description: One or more modalities that this application belongs to
          type: array
          items:
            $ref: '#/components/schemas/Value'
          example: [
            { id: 430e8400-e29b-41d4-a716-446655440000, name: CT},
            { id: 650e8400-e29b-41d4-a716-446655440000, name: MR}]
        category:
          allOf:
            - $ref: '#/components/schemas/Value'
          example: { id: 430e8400-e29b-41d4-a716-446655440000, name : BUSINESSAPP }
    Value:
      type: object
      properties:
        id:
          description: Unique Identifier that represents a lookup value
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          description: Display name of the lookup value
          type: string
          minLength: 3
          maxLength: 2000
          example: Oncology
    Approval:
      type: object
      properties:
        approver:
          description: Email address of the approver of the application
          type: string
          minLength: 3
          maxLength: 255
          example: approver@philips.com
        publishedDateTime:
          type: string
          format: date-time UTC format
          description: Date & Time at which the application was published by the L2 approver
          example: 2024-09-05T17:59:35Z
        propositions:
          description: One or more propositions that this application is approved for
          type: array
          items:
            $ref: '#/components/schemas/Proposition'
        l2comments:
          description: Reason for approval / rejection
          type: string
          minLength: 3
          maxLength: 255
          example: Provide Prod URL instead of CT URL
        markets:
          description: One or more markets that this application is rolled out for
          type: array
          items:
            $ref: '#/components/schemas/Market'
    Ownership:
      type: object
      properties:
        costCenter:
          description: Cost center owning the application cost
          type: string
          minLength: 3
          maxLength: 50
          example: IN5CHS0703
        l1comments:
          description: Reason for L1 approval / rejection
          type: string
          minLength: 3
          maxLength: 255
          example: Provide Prod URL instead of CT URL
    Deployment:
      type: object
      properties:
        url:
          description: Url of the application
          type: string
          minLength: 3
          maxLength: 255
          format: url
          example: https://www.usa.philips.com/healthcare/resources/landing/dosewise-offboarding
        icon:
          description: Icon of the application
          type: string
          minLength: 3
          maxLength: 2000
          format: url
          example: https://d1u65tovhg2ye9.cloudfront.net/icons/doswiseportal.png
        mode:
          description: Application deployment mode
          allOf:
            - $ref: '#/components/schemas/Value'
          example: { id: 430e8400-e29b-41d4-a716-446655440000, name : INSTANCE }
    Application:
      type: object
      properties:
        registration:
          $ref: '#/components/schemas/Registration'
        deployment:
          $ref: '#/components/schemas/Deployment'
        ownership:
          $ref: '#/components/schemas/Ownership'
        approval:
          $ref: '#/components/schemas/Approval'
        status:
          allOf:
            - $ref: '#/components/schemas/Value'
          example: { id: 430e8400-e29b-41d4-a716-446655440000, name : DRAFT }
        lastModifiedDateTime:
          type: string
          format: date-time
          description: Date & Time at which the application was last modified by requestor
          example: 2024-01-22T13:09:04+0000
        enabledForOrg:
          description: Indicates if the Organization (that the currently logged in user is part of) has access to the application
          type: boolean
          example: true
        enabledForUser:
          description: Indicates if the current logged in user has access to the application
          type: boolean
          example: true
        workflowParticipants:
          $ref: '#/components/schemas/GroupList'
...
    ApplicationList:
      type: array
      items:
        $ref: "#/components/schemas/Application"
    ApplicationID:
      type: object
      properties:
        id:
          description: Unique Identifier that represents an application in OneSpace platform
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
...
    Proposition:
      type: object
      required:
        - name
      properties:
        id:
          description: Unique Identifier that represents a proposition in OneSpace platform
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          description: Name of the proposition
          type: string
          minLength: 3
          maxLength: 255
          example: L0 Service Management pack
        description:
          description: Description of the application
          type: string
          minLength: 3
          maxLength: 400
          example: Managed Technology Services is a long term strategic partnership contract that involves technology assessment and planning to identify and track the performance of healthcare technology over the term of the contract. This includes equipment replacements, upgrades, KPI selection, monitoring and reporting
        isActive:
          description: Boolean indicator of whether or not the Proposition is currently active.
          type: boolean
          example: true
        applications:
          type: array
          items:
            type: string
          example: [ "1f76732b-861b-4997-b1f4-788eab3c0cba",
                     "7583f5f4-ea78-442a-99f4-0caddd45fb19"]
...
    ErrorResponse:
      type: object
      properties:
        message:
          description: Error Description message
          type: string
          example: Error while processing Request
        code:
          description: Error code
          type: string
          example: error_code
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: error_category
    BadRequestErrorResponse:
      type: object
      properties:
        message:
          description: Bad Request Error Description
          type: string
          example: Invalid input data provided.
        code:
          description: Error code
          type: string
          example: bad_request
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: client error
    BadRequestError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/BadRequestErrorResponse"
    UnAuthorizedErrorResponse:
      type: object
      properties:
        message:
          description: UnAuthorized Operation Error Description
          type: string
          example: UnAuthorized Access
        code:
          description: Error code
          type: string
          example: invalid_token
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: invalid_token
    UnAuthorizedRequestError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/UnAuthorizedErrorResponse"
...
    ConflictErrorResponse:
      type: object
      properties:
        message:
          description: Error Message for Conflicting Information
          type: string
          example: Resource already exists or Conflict with current state of resource.
        code:
          description: Error code
          type: string
          example: duplicate_key
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: duplicate_key
    ConflictError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/ConflictErrorResponse"
    InternalServerErrorResponse:
      type: object
      properties:
        message:
          description: Error Message for Internal Server Error
          type: string
          example: The server encountered an unexpected condition which prevented it from fulfilling the request.
        code:
          description: Error code
          type: string
          example: database_error
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: database_error
    InternalServerError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/InternalServerErrorResponse"
    Error:
      type: object
      properties:
        issue:
          description: A detailed human-readable message
          type: array
          items:
            $ref: "#/components/schemas/ErrorResponse"
...
    Market:
      type: object
      required:
        - name
      properties:
        id:
          description: Unique Identifier that represents a Market ID
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          description: Name of the market
          type: string
          minLength: 3
          maxLength: 255
          example: NAM
        country:
          description: country of the market
          type: string
          minLength: 3
          maxLength: 255
          example: United States
        region:
          description: Region of the market
          type: string
          minLength: 3
          maxLength: 255
          example: North America
        subRegion:
          description: Sub Region of the market
          type: string
          minLength: 3
          maxLength: 255
          example: North
...
    Group:
      type: object
      properties:
        id:
          description: Unique Identifier that represents an group id
          type: string
          format: uuid
          example: 120d45fc-314a-4d87-831d-d69f253263f6
        name:
          description: Group name
          type: string
          example: ApplicationIntegratorsGroup
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          description: Unique Identifier that represents an user id
          type: string
          format: uuid
          example: 120d45fc-314a-4d87-831d-d69f253263f6
        userName:
          description: User name
          type: string
          example: application.integrator@philips.com
        userEmail:
          description: User email
          type: string
          example: application.integrator@philips.com
    GroupList:
      type: array
      items:
        $ref: "#/components/schemas/Group"
...
  responses:
...
    400:
      description: |
        **Bad Request**. Request is unparsable, syntactically incorrect, or violates 
        schema.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'
    401:
      description: |
        **Unauthorized**. Authorization failure due to missing or invalid session cookie.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/UnAuthorizedRequestError'
    403:
      description: |
        **Forbidden**. Operation is not permitted based on the supplied authorization.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/Error'
...
    405:
      description: |
        **Method Not Allowed**. The method specified for the request is not allowed.
    406:
      description: |
        **Not Acceptable**. The media type is invalid.
    409:
      description: |
        **Conflict**. Request server refused to create or update a resource that already exists or has conflicting information.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/ConflictError'
    413:
      description: |
        **Request Entity Too Large**. The server is refusing to process a request because the 
        request entity is larger than the server is willing or able to process
    415:
      description: |
        **Unsupported Media Type**. The media type in the request is not supported.
    429:
      description: |
        **Too Many Requests**. The request quota has been filled. The server is throttling requests to prevent overload. 
    500:
      description: |
        **Internal Server Error**. The server encountered an unexpected condition which prevented 
        it from fulfilling the request.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/InternalServerError'
    502:
      description: |
        **Bad Gateway**. The server received an invalid response from the upstream server.
    503:
      description: |
        **Service Unavailable**. The server is currently unable to handle the request due to a 
        temporary overloading or maintenance of the server.
    504:
      description: |
        **Gateway Timeout**. The server did not receive a timely response from the upstream 
        server specified it needed to access in attempting to complete the request 
...
  securitySchemes:
    onespaceCookieAuth:
      type: apiKey
      in: cookie
      name: onespacecookie
    bearerAuth:
      type: http
      scheme: bearer
security:
  - onespaceCookieAuth: []
  - bearerAuth: []